requirements.txt
anthropic
google-genai
fastmcp
python-dotenv
httpx
google-search-results
